wip: standardise colour objects and update square btn
