+++ src/components/elements/buttons/button.types.ts
@@ -1,4 +1,5 @@
 import { ButtonHTMLAttributes, ReactNode } from 'react';
+import { ThemeColourGroup } from '@app/theme/palettes/colors.ts';
 
 export type ButtonVariant = 'primary' | 'secondary' | 'outlined' | 'gradient';
 export type ButtonSize = 'xs' | 'small' | 'medium' | 'large';
@@ -18,4 +19,19 @@ export interface ButtonStyleProps {
     $variant?: CommonButtonProps['variant'];
     $size?: CommonButtonProps['size'];
     $color?: CommonButtonProps['color'];
+    $iconPosition?: CommonButtonProps['iconPosition'];
+}
+
+export interface ExtendedButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+    color?: ThemeColourGroup;
+    variant?: ButtonVariant;
+    size?: ButtonSize;
+    colorIntensity?: number;
+}
+
+export interface ExtendedButtonStyleProps {
+    $color: ThemeColourGroup;
+    $variant?: ExtendedButtonProps['variant'];
+    $size?: ExtendedButtonProps['size'];
+    $colorIntensity?: ExtendedButtonProps['colorIntensity'];
 }
