+++ src/components/elements/dialog/Dialog.styles.ts
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { FloatingOverlay } from '@floating-ui/react';
+import { colorsAll } from '@app/theme/palettes/colors.ts';
 
 export const Wrapper = styled.div`
     display: flex;
@@ -22,6 +23,6 @@ export const Overlay = styled(FloatingOverlay)`
     display: flex;
     align-items: center;
     justify-content: center;
-    background-color: ${({ theme }) => theme.colors.darkAlpha[50]};
+    background-color: ${colorsAll.darkAlpha[50]};
     z-index: 2;
 `;
