+++ src/store/appStateStore.ts
@@ -1,4 +1,9 @@
-import { ApplicationsVersions, ExternalDependency } from '@app/types/app-status';
+import {
+    ApplicationsVersions,
+    ExternalDependency,
+    ExternalDependencyStatus,
+    Manufacturer,
+} from '@app/types/app-status';
 import { create } from './create';
 import { invoke } from '@tauri-apps/api';
 import { useAppConfigStore } from './useAppConfigStore';
